+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 203
+
+* Update networking refresher doc
+
+--------------------------------------------------------------------------------
+
 Version 202
 
 * Use cxxstd instead of cxxflags